It works... but why.. I do not know

+++ PlanetGen/Assets/UI Scripts/SaveSystem.cs
@@ -30,7 +30,7 @@ public static class SaveSystem
     public class PlanetStruct
     {
         public string name;
-        public float resolution;
+        public int resolution;
         public float radius;
 
         public int noiseLayersLength;
